Restructured views files
