+++ portal/models/movie.py
@@ -60,10 +60,8 @@ def receive_before_insert(mapper, connection, target):
 @event.listens_for(Movie, 'after_update')
 def receive_after_update(mapper, connection, target):
     if target.status == MovieStatus.MOVIE_STATUS_NEW:
-        print('Queue extract task for movie {}'.format(target.id))
         queue_task("extract_frames", target)
     elif target.status == MovieStatus.MOVIE_STATUS_READY and target.actual_water_level is not None:
-        print('Queue run task for movie {}'.format(target.id))
         queue_task("run", target)
 
 def queue_task(type, movie):
